requirements.txt
streamlit==1.31.1
Pillow==10.2.0
watchdog==3.0.0